chore(Validator): move basic validator to reusable helper
