+++ lib/validate.js
@@ -1,16 +1,8 @@
-import Ajv from 'ajv';
-import addAjvErrors from 'ajv-errors';
+import createAjvInstance from './helper/createAjvInstance';
 
 import schema from '@camunda/element-templates-json-schema/resources/schema.json';
 
-const ajvInstance = new Ajv({
-  allErrors: true,
-  strict: false,
-  code: {
-    source: true
-  }
-});
-addAjvErrors(ajvInstance);
+const ajvInstance = createAjvInstance();
 
 export default ajvInstance.compile(schema);
 export const ajv = ajvInstance;
